Fix NotFound.jsx import path after moving file to pages folder
