+++ frontend/src/pages/NotFound/NotFound.jsx
@@ -1,5 +1,5 @@
 import { Link } from 'react-router-dom';
-import './App.css';
+import "../../App.css";
 
 const NotFound = () => (
   <section className="not-found">
